+++ Sources/KeyboardKit/Gestures/ScrollViewGestureButton.swift
@@ -15,15 +15,14 @@ import SwiftUI
 
  This button can be used within a `ScrollView` since it will
  not block the scroll view's scrolling, despite all gestures
- that is applied to it. The code is so complicated, since it
- is the result of many trials and errors, where every change
- has been tested to not affect the scrolling or any gestures.
+ that is applied to it. The code is complicated, since it is
+ the result of trial and many errors, where every change has
+ been tested to not affect the scrolling or any gestures.
 
  If you don't need to use a scroll view, you should consider
  using a ``GestureButton`` instead. It's way more responsive,
  since it uses a single drag gesture to trigger actions with
- no delay. This is something that this button can't do since
- it has to work in a scroll view.
+ no delay, which however doesn't workin within a scroll view.
 
  Note that the view uses an underlying `ButtonStyle` to make
  gestures work. It can thus not apply another style, but you